requirements.txt
streamlit
pdfplumber
python-docx
spacy
sentence-transformers
docx2txt
PyPDF2
pymongo
openai
pytesseract
matplotlib
numpy
Pillow
reportlab
en-core-web-sm
